Rename package to django-filthyfields v0.1.0a1
- Renamed from django-dirtyfields to django-filthyfields
- Version 0.1.0a1 (alpha release)
- Updated all imports and references
