+++ src/filthyfields/__init__.py
@@ -1,0 +1,8 @@
+"""django-filthyfields library for tracking dirty fields on a Model instance."""
+
+__all__ = ["DirtyFieldsMixin"]
+__version__ = "0.1.0a1"
+
+from filthyfields.filthyfields import DirtyFieldsMixin
+
+VERSION = (0, 1, 0)
